[IMP] - separation of concerns
separation of project actions and project-user actions
